+++ languagetool-core/src/main/java/org/languagetool/language/LanguageIdentifier.java
@@ -20,7 +20,6 @@ package org.languagetool.language;
 
 import com.optimaize.langdetect.LanguageDetector;
 import com.optimaize.langdetect.LanguageDetectorBuilder;
-import com.optimaize.langdetect.cybozu.util.Messages;
 import com.optimaize.langdetect.ngram.NgramExtractors;
 import com.optimaize.langdetect.profiles.LanguageProfile;
 import com.optimaize.langdetect.profiles.LanguageProfileReader;
@@ -244,7 +243,7 @@ public class LanguageIdentifier {
       } catch (Exception e) {
         fasttextEnabled = false;
         RuleLoggerMessage msg = new RuleErrorNotification(this.getClass().getSimpleName(), "-",
-          String.format("Fasttext disabled, failed on '%s': %s", text, ExceptionUtils.getStackTrace(e)));
+          String.format("Fasttext disabled, failed on '%s' (shortText='%s'): %s", text, shortText, ExceptionUtils.getStackTrace(e)));
         RuleLoggerManager.getInstance().log(msg, Level.WARNING);
         fasttextProcess.destroy();
       }
@@ -296,6 +295,19 @@ public class LanguageIdentifier {
       fasttextOut.newLine();
       fasttextOut.flush();
       buffer = fasttextIn.readLine();
+      if (buffer == null) {
+        // hack to see if this helps us debug the rare case of readLine() returning null:
+        try {
+          logger.warn("fasttextIn.readLine() returned null, trying again after short delay for input '" + text + "'");
+          Thread.sleep(10);
+          buffer = fasttextIn.readLine();
+          if (buffer == null) {
+            logger.warn("fasttextIn.readLine() returned null again");
+          }
+        } catch (InterruptedException e) {
+          throw new RuntimeException(e);
+        }
+      }
     }
     String[] values = buffer.split(" ");
     if (values.length % 2 != 0) {