+++ languagetool-core/src/main/java/org/languagetool/rules/spelling/SpellingCheckRule.java
@@ -74,6 +74,7 @@ public abstract class SpellingCheckRule extends Rule {
   private static final String SPELLING_IGNORE_FILE = "/hunspell/ignore.txt";
   private static final String SPELLING_FILE = "/hunspell/spelling.txt";
   private static final String CUSTOM_SPELLING_FILE = "/hunspell/spelling_custom.txt";
+  private static final String GLOBAL_SPELLING_FILE = "spelling_global.txt";
   private static final String SPELLING_PROHIBIT_FILE = "/hunspell/prohibit.txt";
   private static final String CUSTOM_SPELLING_PROHIBIT_FILE = "/hunspell/prohibit_custom.txt";
   private static final String SPELLING_FILE_VARIANT = null;
@@ -209,6 +210,12 @@ public abstract class SpellingCheckRule extends Rule {
 
   @Override
   public abstract RuleMatch[] match(AnalyzedSentence sentence) throws IOException;
+
+  /**
+   * @since 4.8
+   */
+  @Experimental
+  public abstract boolean isMisspelled(String word) throws IOException;
 
   @Override
   public boolean isDictionaryBasedSpellingRule() {
@@ -391,7 +398,8 @@ public abstract class SpellingCheckRule extends Rule {
    * @since 4.8
    */
   public List<String> getAdditionalSpellingFileNames() {
-    return Arrays.asList(language.getShortCode() + CUSTOM_SPELLING_FILE);
+    // NOTE: also add to GermanSpellerRule.getSpeller() when adding items here:
+    return Arrays.asList(language.getShortCode() + CUSTOM_SPELLING_FILE, GLOBAL_SPELLING_FILE);
   }
 
   /**