+++ languagetool-standalone/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 4.8-SNAPSHOT (release planned for 2019-12-27)
 
+#### Catalan
+  * added and improved rules
+
 #### Chinese
   * Now using https://github.com/hankcs/HanLP for tokenization (PR 1981)
 
@@ -44,6 +47,8 @@
   * The unmaintained code from package `org.languagetool.dev.wikipedia.atom`
     has been removed. It hadn't been maintained for years and didn't work properly
     anymore.
+  * `spelling_global.txt` has been added. Words or phrases added here will
+    be accepted for all languages.
   * `prohibit_custom.txt` and `spelling_custom.txt` can be used to make your
     own additions to `spelling.txt` and `prohibit.txt` without having to edit those
     files after a LanguageTool update (you will still need to manually copy those